+++ src/CollapsibleSection/collapsibleSection.scss
@@ -6,7 +6,7 @@ $header-background-selected: transparent !default;
 $header-background-hover: $brc-gradientDarkGrey !default;
 $border-color: $brc-silver !default;
 
-.collapsable-section {
+.collapsible-section {
   border: 1px solid $border-color;
   background: $background;
   cursor: pointer;
