+++ src/CollapsibleSection/examples/collapsibleSection.scss
@@ -1,6 +1,6 @@
 @import '../../utils/theme/variables.scss';
 
-.live-demo .collapsable-section {
+.live-demo .collapsible-section {
   overflow: hidden;
   margin-bottom: 10px;
   border-radius: 4px;
