+++ src/KitchenSink/style.scss
@@ -192,7 +192,7 @@
         }
       }
 
-      .collapsable-section,
+      .collapsible-section,
       .single-item-section {
         border: 0;
 
@@ -239,7 +239,7 @@
       }
 
       .single-item-section,
-      .collapsable-section .content {
+      .collapsible-section .content {
         padding: 20px;
         border-bottom: 1px solid #eaeaea;
         width: 100%;
